update: add auth middleware interface and update firebase middleware implementation
